+++ src/hooks/use-sys-stats.ts
@@ -11,7 +11,7 @@ export function useSystemStats<
      )
      useEffect(()=>{
           const interval = setInterval(async()=>{
-               const stats: SystemStats = await invoke("get_sys_stats");
+               const stats = await invoke<SystemStats>("get_sys_stats");
                const {cpu_usage,cpu_frequency} = stats;
                const avg_usage = cpu_usage.reduce((a,b)=>a+b,0)/cpu_usage.length;
                const avg_freq = cpu_frequency.reduce((a,b)=>a+b,0)/cpu_frequency.length;